[Refactor] Add Utils Repository
