+++ kalliope/core/ConfigurationManager/SettingLoader.py
@@ -1,13 +1,13 @@
 import logging
 
 from YAMLLoader import YAMLLoader
-from kalliope.core.FileManager import FileManager
 from kalliope.core.Models import Singleton
 from kalliope.core.Models.RestAPI import RestAPI
 from kalliope.core.Models.Settings import Settings
 from kalliope.core.Models.Stt import Stt
 from kalliope.core.Models.Trigger import Trigger
 from kalliope.core.Models.Tts import Tts
+from kalliope.core.Utils.FileManager import FileManager
 
 FILE_NAME = "settings.yml"
