+++ kalliope/core/Utils/FileManager.py
