+++ kalliope/core/Utils/__init__.py
@@ -1,0 +1,2 @@
+from kalliope.core.Utils import Utils
+from kalliope.core.Utils import FileManager
