+++ kalliope/core/__init__.py
@@ -1,6 +1,7 @@
 from kalliope.core.OrderAnalyser import OrderAnalyser
 from kalliope.core.OrderListener import OrderListener
 from kalliope.core.ShellGui import ShellGui
-from kalliope.core.FileManager import FileManager
 from kalliope.core.Utils import Utils
+from kalliope.core.Utils import FileManager
+
